refactor: data Userクラスのファイルを独立したファイルに移動
